feat(ERASMUS-122): added pre approval page for students. added routes to main.tsx file. fixed user protected route logic.
