+++ frontend/src/components/Layout.tsx
@@ -15,9 +15,6 @@ const Layout = () => {
                     main: { backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[8] : theme.colors.gray[0] },
                 })}
             >
-                <Group position="center">
-                    <Title>Welcome to Beam</Title>
-                </Group>
                 <Outlet />
             </AppShell>
         </main>
