+++ frontend/src/components/auth/RequireAuth.tsx
@@ -10,7 +10,7 @@ const RequireAuth = ({ allowedUsers }: RequireAuthProps) => {
     const { user } = useUser()
     const location = useLocation()
 
-    const isUser = (x: any): x is User => Object.keys(x).length === 0
+    const isUser = (x: any): x is User => Object.keys(x).length !== 0
 
     // User did not login at all
     if (!isUser(user)) {
